Extract ConstantExpressionVerifier
The functionality doesn't really belong in ExpressionInterpreter, which is already
doing too much.
